Refactor: extract correlation atoms into a dedicated file
Extracted correlation-related Jotai atoms (`correlationAlphaAtom`, `correlationAlternativeAtom`, and `correlationMethodAtom`) from `src/atom/dataAtom.ts` into a new `src/atom/correlationAtom.ts` file to improve state organization and maintainability. Updated the relevant imports in dependent layout components. No core logic or features were changed.
