+++ src/atom/dataAtom.ts
@@ -76,10 +76,6 @@ export const aiKeyAtom = atomWithStorage<string | null>("key", null);
 export const aiModelAtom = atomWithStorage<string>("model", "gemini-3-flash-preview");
 export const gistTokenAtom = atomWithStorage<string | null>("gistToken", null);
 
-export const correlationAlphaAtom = atomWithStorage<number>("correlationAlpha", 0.01);
-export const correlationAlternativeAtom = atomWithStorage<"two-sided" | "less" | "greater">("correlationAlternative", "two-sided");
-export const correlationMethodAtom = atomWithStorage<"spearman" | "pearson">("correlationMethod", "spearman");
-
 export const nonInferredDataAtom = atom((get) => {
   const data = get(dataAtom);
   return data.filter((item) => !item[3]?.inferred);
